services/payment_service/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-jaeger
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
requests